PI-2519 Add test for address matching
Note: this currently fails intermittently due to Logstash caching (see PI-2519).
